[refactor] move BVH to Math/Algorithm

+++ Engine/Engine/Source/Math/Algorithm/IndexedKdtree/IndexedKdtreeParams.h
@@ -11,14 +11,14 @@ public:
 	IndexedKdtreeParams();
 
 	std::size_t getMaxNodeItems() const;
-	int getTraversalCost() const;
-	int getInteractCost() const;
+	float getTraversalCost() const;
+	float getInteractCost() const;
 	float getEmptyBonus() const;
 
 private:
 	std::size_t m_maxNodeItems;
-	int         m_traversalCost;
-	int         m_interactCost;
+	float       m_traversalCost;
+	float       m_interactCost;
 	float       m_emptyBonus;
 };
 
@@ -26,8 +26,8 @@ private:
 
 inline IndexedKdtreeParams::IndexedKdtreeParams() : 
 	m_maxNodeItems (1),
-	m_traversalCost(1),
-	m_interactCost (80),
+	m_traversalCost(1.0f),
+	m_interactCost (80.0f),
 	m_emptyBonus   (0.5f)
 {}
 
@@ -36,12 +36,12 @@ inline std::size_t IndexedKdtreeParams::getMaxNodeItems() const
 	return m_maxNodeItems;
 }
 
-inline int IndexedKdtreeParams::getTraversalCost() const
+inline float IndexedKdtreeParams::getTraversalCost() const
 {
 	return m_traversalCost;
 }
 
-inline int IndexedKdtreeParams::getInteractCost() const
+inline float IndexedKdtreeParams::getInteractCost() const
 {
 	return m_interactCost;
 }
